+++ examples/sqlalchemy_example.py
@@ -1,6 +1,6 @@
 from sqlalchemy import create_engine, select, text
 
-from duckdb_engine import URL, read_parquet
+from duckdb_sqlalchemy import URL, read_parquet
 
 engine = create_engine(URL(database=":memory:"))